#157 - Move calculate_proximity -> Trajectory
