+++ ecoscope/base/base.py
@@ -1,5 +1,9 @@
+import itertools
+import uuid
 import warnings
 from functools import cached_property
+from dataclasses import dataclass, field
+from typing import Any, List
 
 import geopandas as gpd
 import numpy as np
@@ -14,6 +18,22 @@ from ecoscope.base._dataclasses import (
     RelocsSpeedFilter,
     TrajSegFilter,
 )
+
+
+@dataclass
+class SpatialFeature:
+    """
+    A spatial geometry with an associated name and unique ID. Becomes a useful construct in several movdata calculations
+    """
+
+    name: str = ""
+    unique_id: Any = uuid.uuid4()
+    geometry: Any = None
+
+
+@dataclass
+class ProximityProfile:
+    spatial_features: List[SpatialFeature] = field(default=list)
 
 
 class EcoDataFrame(gpd.GeoDataFrame):
@@ -653,3 +673,40 @@ class Trajectory(EcoDataFrame):
 
         instance = Properties()
         return instance
+
+    def calculate_proximity(
+        self,
+        proximity_profile: ProximityProfile,
+    ) -> pd.DataFrame:
+        """
+        A function to analyze the trajectory of a subject in relation to a set of spatial features and regions to
+        determine where/when the subject was proximal to the spatial feature.
+
+        Parameters
+        ----------
+        proximity_profile: ProximityProfile
+            proximity setting for performing calculation
+        Returns
+        -------
+        pd.DataFrame
+
+        """
+        proximity_events = []
+
+        def analysis(traj):
+            for sf in proximity_profile.spatial_features:
+                proximity_dist = traj.geometry.distance(sf.geometry)
+                start_fix = gpd.GeoSeries([shapely.Point(g.coords[0]) for g in traj.geometry])
+
+                pr = traj[["groupby_col", "speed_kmhr", "heading"]]
+                pr["proximity_distance"] = proximity_dist
+                pr["proximal_fix"] = start_fix  # TODO: figure out the estimated fix interpolated along the seg
+                pr["estimated_time"] = traj.segment_start
+                pr["geometry"] = traj.geometry
+                pr["spatialfeature_id"] = list(itertools.repeat(sf.unique_id, pr.shape[0]))
+                pr["spatialfeature_name"] = list(itertools.repeat(sf.name, pr.shape[0]))
+
+                proximity_events.append(pr)
+
+        self.groupby("groupby_col")[self.columns].apply(analysis, include_groups=False)
+        return pd.concat(proximity_events).reset_index(drop=True)
